requirements.txt
py-cord
aiosqlite
python-dotenv
easy_pil
python-dateutil